+++ src/utils/formatNumber.js
@@ -1,0 +1,1 @@
+fn.formatNumber = (number, options) => Number(number).toLocaleString(null, options);